+++ examples/main/shared/editor/EditorPlugins.jsx
@@ -221,7 +221,7 @@ const videoHandlers = {
       updateEntity({ data: videoToUpload });
       //updateEntity({ error: { msg: 'Upload Failed' } });
       console.log('consumer uploaded ', videoToUpload);
-    }, 500);
+    }, 1500);
   },
 };