+++ examples/main/shared/viewer/Viewer.jsx
@@ -7,6 +7,10 @@ import * as Plugins from './ViewerPlugins';
 import theme from '../theme/theme'; // must import after custom styles
 import getImagesData from 'wix-rich-content-fullscreen/src/lib/getImagesData';
 import Fullscreen from 'wix-rich-content-fullscreen';
+import {
+  TextSelectionListener,
+  ViewerInlineToolBar,
+} from 'wix-rich-content-text-selection-toolbar';
 import { GALLERY_TYPE } from 'wix-rich-content-plugin-gallery';
 const anchorTarget = '_top';
 const relValue = 'noreferrer';
@@ -59,7 +63,7 @@ export default class Viewer extends PureComponent {
       seoMode,
     };
 
-    return (
+    return [
       <div id="rich-content-viewer" className="viewer">
         <RichContentViewer
           helpers={this.helpers}
@@ -78,8 +82,11 @@ export default class Viewer extends PureComponent {
             index={expandModeIndex}
           />
         )}
-      </div>
-    );
+      </div>,
+      !isSSR() && (
+        <TextSelectionListener targetId={'rich-content-viewer'} ToolBar={ViewerInlineToolBar} />
+      ),
+    ];
   }
 }