+++ packages/editor-common/web/src/Base/createBaseInsertPluginButton.jsx
@@ -33,7 +33,10 @@ export default ({
       const { buttonStyles } = props.theme || {};
       this.styles = mergeStyles({ styles, theme: buttonStyles });
       this.buttonRef = React.createRef();
+      this.toolbarName = props.toolbarName;
     }
+
+    onPluginAdd = name => helpers?.onPluginAdd?.(blockType, name || this.toolbarName);
 
     componentDidMount() {
       this.initialIntent();
@@ -63,6 +66,7 @@ export default ({
     };
 
     createBlock = (editorState, data, type) => {
+      this.onPluginAdd();
       this.props.hidePopup?.();
       return createBlock(editorState, data, type);
     };
@@ -94,10 +98,12 @@ export default ({
           this.toggleButtonModal(event);
           break;
         case 'custom-block':
+          this.onPluginAdd(name);
           this.addCustomBlock(button);
           break;
         default:
           this.addBlock(button.componentData || {});
+          break;
       }
     };
 
@@ -280,6 +286,7 @@ export default ({
     isMobile: PropTypes.bool,
     t: PropTypes.func,
     tabIndex: PropTypes.number,
+    toolbarName: PropTypes.string,
   };
 
   return InsertPluginButton;