+++ CHANGELOG.md
@@ -12,14 +12,19 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
+  
+### :bug: Bug Fix
+- `video`
+  - [#746](https://github.com/wix-incubator/rich-content/pull/746) controls not showing  
+  
 </details>
 <hr/>
 
 ## 6.9.5 (Mar 4, 2020)
 ### :bug: Bug Fix
 - `editor`
-  - [#745](https://github.com/wix-incubator/rich-content/pull/745) hotfix - editorState trigger
-  - 
+  - [#745](https://github.com/wix-incubator/rich-content/pull/745) editorState trigger
+
 ## 6.9.4 (Mar 4, 2020)
 ### :bug: Bug Fix
 - `html`