+++ CHANGELOG.md
@@ -13,9 +13,58 @@
     Click to see more.
   </summary>
 
-
+### :bug: Bug Fix
+- `hashtag`
+  - [#707](https://github.com/wix-incubator/rich-content/pull/707) allow hashtag and link in the same block
+  
 </details>
 <hr/>
+
+## 6.9.6 (Mar 5, 2020)
+### :bug: Bug Fix
+- `editor`
+  - [#747](https://github.com/wix-incubator/rich-content/pull/747) unable to build - draftUtils Bi consumption
+- `video`
+  - [#746](https://github.com/wix-incubator/rich-content/pull/746) controls not showing
+
+## 6.9.5 (Mar 4, 2020)
+### :bug: Bug Fix
+- `editor`
+  - [#745](https://github.com/wix-incubator/rich-content/pull/745) editorState trigger
+
+## 6.9.4 (Mar 4, 2020)
+### :bug: Bug Fix
+- `html`
+  - [#732](https://github.com/wix-incubator/rich-content/pull/732) reload issue in ios
+### :rocket: New Feature
+- `bi-events`
+  - [#675](https://github.com/wix-incubator/rich-content/pull/514) supports `onPluginAdd` (toolbar only) and `onPluginDelete` callbacks via props
+### :house: Internal
+- `gallery`
+  - [#742](https://github.com/wix-incubator/rich-content/pull/742) bump pro-gallery version to 1.7.25
+
+## 6.9.3 (Mar 4, 2020)
+### :bug: Bug Fix
+- `html`
+  - [#732](https://github.com/wix-incubator/rich-content/pull/732) fix html width in viewer
+### :house: Internal
+- `e2e`
+  - [#737](https://github.com/wix-incubator/rich-content/pull/737) fix rendering and plugins tests
+
+## 6.9.2 (Feb 27, 2020)
+### :bug: Bug Fix
+- `video`
+  - [#727](https://github.com/wix-incubator/rich-content/pull/727) loader is consistent throughout inital file upload
+- `editor`
+  - [#731](https://github.com/wix-incubator/rich-content/pull/731) tooltips disappear after click
+### :house: Internal
+- `gallery`
+  - [#724](https://github.com/wix-incubator/rich-content/pull/724) bump pro-gallery version to 1.7.23
+
+## 6.9.1 (Feb 25, 2020)
+### :bug: Bug Fix
+- `image`
+  - [#729](https://github.com/wix-incubator/rich-content/pull/729) image with no config
 
 ## 6.9.0 (Feb 25, 2020)
 ### :rocket: New Feature
@@ -30,6 +79,7 @@
   - [#712](https://github.com/wix-incubator/rich-content/pull/712) old image client data support #2
 - `video`
   - [#648](https://github.com/wix-incubator/rich-content/pull/648) loader disappears before video is fully loaded
+
 ### :house: Internal
 - `html`
   - [#681](https://github.com/wix-incubator/rich-content/pull/681) removed htmlIframeSrc from config. Loads iframe html from bundle instead of the static file served by the consumer